Change file-url mixed upload
change the "_partition" method to create one list for file-url mixed input. where type specifies if it is a file or url.
change the "upload_files" and "upload_urls" method. for each file/url the method is called once now instead of one call for all files/urls.
change the "upload_urls" to not download the url before adding it to ptpimg.
change the "_perform" as required after changing "upload_files" and "upload_urls"